refactor(cli): extract workflows command module
